Criado tratamento de exceções

+++ src/main/java/dev/gabrielbarbosa/dscommerce/services/ProductService.java
@@ -3,24 +3,28 @@ package dev.gabrielbarbosa.dscommerce.services;
 import dev.gabrielbarbosa.dscommerce.dto.ProductDTO;
 import dev.gabrielbarbosa.dscommerce.entities.Product;
 import dev.gabrielbarbosa.dscommerce.repositories.ProductRepository;
+import dev.gabrielbarbosa.dscommerce.services.exceptions.DatabaseException;
+import dev.gabrielbarbosa.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class ProductService {
+
+    private final String PRODUCT_NOT_FOUND = "PRODUTO NÃO ENCONTRADO.";
 
     @Autowired
     private ProductRepository productRepository;
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = productRepository.findById(id).get();
+        Product product = productRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(PRODUCT_NOT_FOUND));
         return new ProductDTO(product);
     }
 
@@ -38,13 +42,25 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO) {
-        Product product = productRepository.getReferenceById(id);
-        product.update(productDTO);
-        return new ProductDTO(product);
+        try {
+            Product product = productRepository.getReferenceById(id);
+            product.update(productDTO);
+            return new ProductDTO(product);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException(PRODUCT_NOT_FOUND);
+        }
     }
 
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException(PRODUCT_NOT_FOUND);
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("FALHA DE INTEGRIDADE REFERENCIAL.");
+        }
     }
 
 }
